// resources/js/interfaces/product.ts
import CategoryInterface from "./category";
import UserInterface from "./user";
import ProductInfoInterface from "./product-info";
import TagInterface from "./tag";
import DailyDealInterface from "./daily-deal";
import RateInterface from "./rate";
export default interface ProductInterface {
    id: number;
    user_id: number;
    category_slug: string;
    title: string;
    slug?: string;
    price: number | string;
    priceInt?: number;
    save: number | string;
    youSave?: string | number;
    youSaveInt?: number;
    qty: number;
    cartQty?: number;
    colors: string[];
    sizes: string[];
    images: string[] | any;
    info: string;
    img_path: string;
    saved_price: number | string;
    saved_priceInt?: number;
    rate_avg?: number;
    total?: number | string;
    totalInt?: number;
    owner?: UserInterface;
    category?: CategoryInterface;
    rates?: RateInterface[];
    tags?: TagInterface[];
    pi?: ProductInfoInterface;
    daily?: DailyDealInterface;
    created_at?: string;
    updated_at?: string;
    deleted_at?: string;
    cart_exists?: boolean;
}